# Conflicts:
#	ComplainManager.xcworkspace/xcuserdata/monika.xcuserdatad/UserInterfaceState.xcuserstate
#	ComplainManager/LoginViewController.m
#	ComplainManager/UserService.m

+++ ComplainManager/AddComplainCell.h
@@ -13,7 +13,7 @@
 @property (weak, nonatomic) IBOutlet UIButton *deleteImageButton;
 @property (weak, nonatomic) IBOutlet UIButton *selectImageButton;
 
-- (void)displayData:(int)index data:(NSMutableArray *)imageArray;
+- (void)displayData:(long)index data:(NSMutableArray *)imageArray;
     
 
 @end
